chore: refactor command and metrics handling, introducing CommandFrameworkModule and MetricsModule
